Move fake into separate file
