+++ internal/search/exhaustive/service/BUILD.bazel
@@ -3,7 +3,10 @@ load("//dev:go_defs.bzl", "go_test")
 
 go_library(
     name = "service",
-    srcs = ["search.go"],
+    srcs = [
+        "fake.go",
+        "search.go",
+    ],
     importpath = "github.com/sourcegraph/sourcegraph/internal/search/exhaustive/service",
     visibility = ["//:__subpackages__"],
     deps = [
@@ -14,7 +17,7 @@ go_library(
 
 go_test(
     name = "service_test",
-    srcs = ["search_test.go"],
+    srcs = ["fake_test.go"],
     deps = [
         ":service",
         "@com_github_stretchr_testify//require",
